ADD: Commit for Mayo Server

+++ src/remove_background_slice_label.py
@@ -1,6 +1,6 @@
 import os
 
-root_path = "/media/Datacenter_storage/Ji/valdo_dataset/valdo_t2s_cmbOnly_GAN"
+root_path = "/media/Datacenter_storage/Ji/valdo_dataset/valdo_resample_ALFA_YOLO_PNG_epd_gt_box_t2s_GAN_3slices_cmbTrainOnly"
 images_dir = f"{root_path}/images/train"
 labels_dir = f"{root_path}/labels/train"
 
@@ -12,10 +12,10 @@ for label_file in os.listdir(labels_dir):
             image_file = label_file.replace('.txt', '.png')
             image_path = os.path.join(images_dir, image_file)
             # mask_path = image_path.replace('images', 'masks')
-            os.remove(label_path)
-            os.remove(image_path)
-            # os.remove(mask_path)
-            print(f"Deleted: {label_path} and {image_path}")
+            if not os.path.exists(image_path):
+                os.remove(label_path)
+                # os.remove(mask_path)
+                print(f"Deleted: {label_path} and {image_path}")
 print("Cleanup complete.")
 
 '''
